+++ src/ansibleguy-webui/aw/config/form_metadata.py
@@ -13,7 +13,8 @@ FORM_LABEL = {
             'credentials_needed': 'Needs Credentials',
             'credentials_default': 'Default Job Credentials',
             'credentials_category': 'Credentials Category',
-            'form': 'Execution Form',
+            'execution_prompts_required': 'Execution Prompts - Required',
+            'execution_prompts_optional': 'Execution Prompts - Optional',
         },
         'credentials': {
             'connect_user': 'Connect User',
@@ -40,18 +41,6 @@ FORM_LABEL = {
             'git_hook_post': 'Git Post-Hook',
             'git_override_initialize': 'Git Override Initialize-Command',
             'git_override_update': 'Git Override Update-Command',
-        },
-        'execution_form_field': {
-            'help': 'Help Text',
-            'var': 'Variable',
-            'var_type': 'Variable Type',
-            'field_type': 'Field Type',
-            'multiple': 'Multiple Choices',
-            'separator': 'Multiple-Choice Separator',
-            'validate_error': 'Validation Error-Message',
-            'validate_regex': 'Validation Regex',
-            'validate_max': 'Maximum Value',
-            'validate_min': 'Minimum Value',
         },
     },
     'settings': {
@@ -141,7 +130,10 @@ FORM_HELP = {
             'credentials_category': 'The credential category can be used for dynamic matching of '
                                     'user credentials at execution time',
             'enabled': 'En- or disable the schedule. Can be ignored if no schedule was set',
-            'form': 'Select a Job-Execution form to display on ad-hoc executions',
+            'execution_prompts_required': 'Required job attributes and/or variables to prompt at custom execution. '
+                                          'Comma-separated list of key-value pairs.<br>'
+                                          "Variables can be supplied like so: 'var={VAR-NAME}#{DISPLAY-NAME}'<br>"
+                                          "Example: 'limit,check,var=add_user#User to add' ",
         },
         'credentials': {
             'vault_file': 'Path to the file containing your vault-password',
@@ -172,18 +164,6 @@ FORM_HELP = {
                                        '(clone) the repository',
             'git_override_update': 'Advanced usage! Completely override the command used to update '
                                    '(pull) the repository',
-        },
-        'execution_form_field': {
-            'var_type': '<b>WARNING</b>: Choose env-var for secret values! Commandline arguments are viewable and '
-                        'will be logged',
-            'multiple': 'If multiple the form should support multiple user-choices',
-            'separator': 'If multiple choices are used - they will be joined on serialization before they get passed '
-                         'to the Ansible Execution. This separator will be used on that join operation',
-            'validate_error': 'Error-Message to display on validation error',
-            'validate_regex': 'Regex to validate the value. Test it here: '
-                              '<a href="https://regex101.com/">regex101.com</a>. (flavor needs to be JavaScript)',
-            'validate_max': 'Field-type integer: maximum integer; Field-type string: maximum length',
-            'validate_min': 'Field-type integer: minimum integer; Field-type string: minimum length',
         },
     },
     'settings': {