+++ src/ansibleguy-webui/aw/execute/queue.py
@@ -1,18 +1,17 @@
-from aw.model.job import Job, JobQueue
+from aw.model.job import JobExecution, JobQueue
 from aw.utils.debug import log
-from aw.base import USERS
 
 
-def queue_get() -> (tuple[Job, USERS], None):
+def queue_get() -> (JobExecution, None):
     next_queue_item = JobQueue.objects.order_by('-created').first()
     if next_queue_item is None:
         return None
 
-    job, user = next_queue_item.job, next_queue_item.user
+    execution = next_queue_item.execution
     next_queue_item.delete()
-    return job, user
+    return execution
 
 
-def queue_add(job: Job, user: USERS):
-    log(msg=f"Job '{job.name}' added to execution queue", level=4)
-    JobQueue(job=job, user=user).save()
+def queue_add(execution):
+    log(msg=f"Job '{execution.job.name} {execution.id}' added to execution queue", level=4)
+    JobQueue(execution=execution).save()